Remove resolution based on class name via env var
