+++ jaeger-core/src/test/java/io/jaegertracing/senders/FaultySenderFactory.java
@@ -1,0 +1,29 @@
+/*
+ * Copyright (c) 2018, The Jaeger Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+package io.jaegertracing.senders;
+
+import io.jaegertracing.Configuration;
+
+public class FaultySenderFactory implements SenderFactory {
+  @Override
+  public Sender getSender(Configuration.SenderConfiguration senderConfiguration) {
+    throw new RuntimeException("boo");
+  }
+
+  @Override
+  public String getType() {
+    return "faulty";
+  }
+}
